requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
matplotlib-venn
biopython
gprofiler-official
requests
plotly
networkx
pyvis
scipy
statsmodels
spacy
gseapy
mygene
venny4py
altair
detect_delimiter
stmol